Project 07 and 08: Build a vm translator
Building the VM translator (vmtranslator.py) - a program that translates programs written in the VM language into programs written in the Hack machine language.
Given a Prog.vm, a HACK assembly file Prog.asm should be created, which includes the bootstrap code as well. The file should produce the binary when fed to the assembler.